+++ packages/did-ui-react/src/components/config-provider/types.ts
@@ -1,5 +1,5 @@
 import { IConfig } from '@portkey/types';
-import { ISocialLoginConfig, TCustomNetworkType, TSupportAccountType } from '../../types';
+import { ISocialLoginConfig, NetworkType, TCustomNetworkType, TSupportAccountType } from '../../types';
 import { BaseReCaptcha } from '../types';
 
 export interface ConfigProviderProps {
@@ -18,4 +18,5 @@ export interface GlobalConfigProps extends IConfig {
   serviceUrl?: string;
   loginConfig?: ILoginConfig;
   customNetworkType?: TCustomNetworkType;
+  networkType?: NetworkType;
 }