+++ packages/did-ui-react/src/components/context/PortkeyAssetProvider/actions.ts
@@ -45,6 +45,7 @@ export type BaseAssetProps = {
   originChainId: ChainId;
   managerPrivateKey?: string;
   didStorageKeyName?: string;
+  isLoginOnChain?: boolean;
 };
 
 export type TxFeeType = {