+++ packages/did/src/types.ts
@@ -19,6 +19,7 @@ import {
   TransactionStatus,
   MultiTransactionParamInfo,
   MultiChainInfo,
+  LoginStatusEnum,
 } from '@portkey/types';
 
 export type LoginType = 'scan' | 'account';
@@ -97,6 +98,8 @@ export interface IDIDAccountMethods extends IAccountMethods {
   getHolderInfo(params: Omit<GetHolderInfoParams, 'manager'>): Promise<IHolderInfo>;
   getVerifierServers(chainId: ChainId): Promise<VerifierItem[]>;
   getCAHolderInfo(originChainId: ChainId): Promise<CAHolderInfo>;
+  saveTempStatus(params: { chainId: ChainId; caHash: string; caAddress: string; sessionId: string }): void;
+  updateLoginStatus(params: LoginStatusEnum): void;
   /**
    * @param caHash - CA wallet hash
    * @param managementAddress - address of managementAccount